backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
deprecation==2.1.0
ecdsa==0.19.0
fastapi==0.115.5
frozenlist==1.5.0
gotrue==2.10.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
Jinja2==3.1.4
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.2.1
packaging==23.1
postgrest==0.18.0
propcache==0.2.0
pyasn1==0.6.1
pydantic==2.10.1
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
realtime==2.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
storage3==0.9.0
supabase==2.10.0
supafunc==0.7.0
typing_extensions==4.12.2
uuid
uvicorn==0.32.1
websockets==13.1
yarl==1.18.0
